Refactor UnicodeText to support ICU on Windows
* Cleans up Chakra.Runtime.PlatformAgnostic vcxprojs
* Converts icu::Normalizer2 to UNormalizer2
* Creates new "POSIX" platform for code that is common across Linux and Unix but not Windows
